xong 100% resetpassword

+++ src/main/java/com/isp/project/repositories/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.isp.project.model.Employee;
 
-public interface UserRepository extends JpaRepository<Employee, Long> {
+public interface UserRepository extends JpaRepository<Employee, Integer> {
     Employee findByUserName(String username);
 
     Employee findByEmail(String email);
